+++ dbus-serialbattery/standalone_serialbattery.py
@@ -55,6 +55,7 @@
 # 1 : JKBMS bluettooth - Jkbms_Ble
 # 2 : JDB bluetooth    - LltJbd_Ble
 # 3 : LiTime bluetooth - LiTime_Ble
+# 4 : Kilovault bluetooth - Kilovault_Ble
 # 10: CAN devices for JKBAMS and DALY
 #
 # devadr
@@ -147,7 +148,6 @@ if "Sinowealth" in BMS_TYPE:
 
 
 class standalone_serialbattery:
-
     def init_bms_types(self):
         self.supported_bms_types = supported_bms_types
         self.expected_bms_types = [battery_type for battery_type in self.supported_bms_types if battery_type["bms"].__name__ in BMS_TYPE or len(BMS_TYPE) == 0]
@@ -271,7 +271,7 @@ class standalone_serialbattery:
             Import ble classes only, if it's a ble port, else the driver won't start due to missing python modules
             This prevent problems when using the driver only with a serial connection
             """
-            if self.driveroption <= 3:  # bluetooth
+            if self.driveroption <= 4:  # bluetooth
 
                 if self.driveroption == 1:  # "Jkbms_Ble":
                     # noqa: F401 --> ignore flake "imported but unused" error
@@ -284,6 +284,10 @@ class standalone_serialbattery:
                 if self.driveroption == 3:  # "LiTime_Ble":
                     # noqa: F401 --> ignore flake "imported but unused" error
                     from bms.litime_ble import LiTime_Ble  # noqa: F401
+
+                if self.driveroption == 4:  # "Kilovault_Ble":
+                    # noqa: F401 --> ignore flake "imported but unused" error
+                    from bms.kilovault_ble import Kilovault_Ble  # noqa: F401
 
                 class_ = eval(self.devpath)
                 testbms = class_("ble_" + self.devadr.replace(":", "").lower(), 9600, self.devadr)
@@ -471,6 +475,15 @@ class standalone_serialbattery:
             Status.append(self.soc)
 
         except Exception as e:
+            (
+                exception_type,
+                exception_object,
+                exception_traceback,
+            ) = sys.exc_info()
+            file = exception_traceback.tb_frame.f_code.co_filename
+            line = exception_traceback.tb_lineno
+            logging.error("Non blocking exception occurred: " + f"{repr(exception_object)} of type {exception_type} in {file} line #{line}")
+            # Ignore any malfunction test_function()
             logging.error("Error during reading BMS")
             logging.error(str(e))